Moved functions and classes to pytorch_utils.py
