+++ src/transformers/models/mmbt/modeling_mmbt.py
@@ -21,7 +21,7 @@ from torch import nn
 from torch.nn import CrossEntropyLoss, MSELoss
 
 from ...modeling_outputs import BaseModelOutputWithPooling, SequenceClassifierOutput
-from ...modeling_utils import ModuleUtilsMixin
+from ...pytorch_utils import ModuleUtilsMixin
 from ...utils import add_start_docstrings, add_start_docstrings_to_model_forward, logging, replace_return_docstrings
