+++ src/transformers/models/speech_to_text/modeling_tf_speech_to_text.py
@@ -20,6 +20,8 @@ from typing import Dict, Optional, Tuple
 
 import tensorflow as tf
 
+import transformers.pytorch_utils
+
 from ...activations_tf import get_tf_activation, glu
 from ...modeling_tf_outputs import (
     TFBaseModelOutput,
@@ -130,7 +132,7 @@ class TFConv1dSubsampler(tf.keras.layers.Layer):
         self.kernel_sizes = config.conv_kernel_sizes
 
         self.conv_layers = [
-            tf.keras.layers.Conv1D(
+            transformers.pytorch_utils.Conv1D(
                 filters=self.mid_channels if i < self.num_layers - 1 else self.out_channels * 2,
                 kernel_size=k,
                 strides=2,
